+++ private_values/main.py
@@ -1,5 +1,6 @@
 import pathlib
 from typing import *
+import abc
 
 
 # =====================================================================================================================
@@ -8,6 +9,7 @@ Type_Path = Union[str, pathlib.Path]
 Type_Value = Union[str, NoReturn, None]
 
 
+# =====================================================================================================================
 class Exx_PvNotAccepted(Exception):
     """
     Any final exception when value can't be get.
@@ -16,11 +18,66 @@ class Exx_PvNotAccepted(Exception):
 
 
 # =====================================================================================================================
+class PrivateAuth:
+    USER: str
+    PWD: str
+
+
+class PrivateTgBotAddress:
+    LINK_ID: str     # @mybot20230913
+    NAME: str        # MyBotPublicName
+    TOKEN: str
+
+
+# =====================================================================================================================
 class PrivateBase:
     RAISE_EXX: bool = True
 
+    def create_attributes(self, attrs: Dict[str, Any]) -> None:
+        for key, value in attrs.items():
+            setattr(self, key, value)
+        self.check_annotations()
 
-class PrivateBaseWFile(PrivateBase):
+    def check_annotations(self) -> Optional[NoReturn]:
+        # print(self.__class__.__mro__)
+        annots = set()
+        for cls in self.__class__.__mro__[:-1]:
+            annots.update(set(cls.__annotations__))
+        # print(annots)
+
+        for attr in annots:
+            # apply losercase for INI
+            if not hasattr(self, attr) and hasattr(self, attr.lower()):
+                setattr(self, attr, getattr(self, attr.lower()))
+
+            if not hasattr(self, attr):
+                msg = f"[CRITICAL]no[{attr=}]"
+                if attr.lower() != attr:
+                    msg += f"\nif used Ini - use only LOWERCASE attrs!!!"
+                msg += f"\n{self.__class__.__name__}"
+                msg += f"\nall your annotations={annots}"
+                raise Exx_PvNotAccepted(msg)
+
+
+# =====================================================================================================================
+class _PrivateBaseWFile_Interface(abc.ABC):
+    @abc.abstractmethod
+    def _get_value_unsafe(self, name: str, section: str, text: str) -> Optional[str]:
+        # NOTICE!!!! "section: Optional[str] = None" - dont code like this!!!
+        # the method work always UNSAFE! any param can't be NONE!
+        pass
+
+    @abc.abstractmethod
+    def _get_section_unsafe(self, section: str, text: str) -> Optional[Dict[str, Any]]:
+        """
+        return
+            NONE - if no section! dont raise inside!
+            {} - if no names!
+        """
+        pass
+
+
+class PrivateBaseWFile(PrivateBase, _PrivateBaseWFile_Interface):
     SECTION: str = None
 
     DIRPATH: Type_Path = pathlib.Path.home()
@@ -47,6 +104,8 @@ class PrivateBaseWFile(PrivateBase):
         else:
             self.DIRPATH = pathlib.Path(_filepath).parent
             self.FILENAME = pathlib.Path(_filepath).name
+
+        self.get_section()
 
     @property
     def filepath(self) -> pathlib.Path:
@@ -95,11 +154,6 @@ class PrivateBaseWFile(PrivateBase):
             print(msg)
             return
 
-    def _get_value_unsafe(self, name: str, section: str, text: str) -> Optional[str]:
-        # NOTICE!!!! "section: Optional[str] = None" - dont code like this!!!
-        # the method work always UNSAFE! any param can't be NONE!
-        raise NotImplementedError
-
     # -----------------------------------------------------------------------------------------------------------------
     def get_section(
             self,
@@ -108,7 +162,7 @@ class PrivateBaseWFile(PrivateBase):
             _filename: str = None,
             _filepath: Type_Path = None,
             _raise_exx: Optional[bool] = None
-    ) -> Optional['PrivateBaseWFile']:
+    ) -> Union['PrivateBaseWFile', NoReturn, None]:
         if _raise_exx is None:
             _raise_exx = self.RAISE_EXX
 
@@ -129,9 +183,8 @@ class PrivateBaseWFile(PrivateBase):
         filetext = _filepath.read_text()
 
         section_dict = self._get_section_unsafe(section=_section, text=filetext)
-        if section_dict is not None:
-            for key, value in section_dict.items():
-                setattr(self, key, value)
+        if section_dict:
+            self.create_attributes(section_dict)
             return self
 
         msg = f"[CRITICAL]no {_section=} in {_filepath=}!"
@@ -143,14 +196,6 @@ class PrivateBaseWFile(PrivateBase):
         else:
             print(msg)
             return
-
-    def _get_section_unsafe(self, section: str, text: str) -> Optional[Dict[str, Any]]:
-        """
-        return
-            None if NO section
-            {} - if no names!
-        """
-        raise NotImplementedError
 
 
 # =====================================================================================================================